DirectFcnCall.Emit stack savings (allows longer call chains)
Originally checked in by DEVSENSE\Jakub on 2012-05-27 20:56:13.62 as Changeset 2947

git-tfs-id: [https://devsense.visualstudio.com/DefaultCollection]$/Phalanger/Main;C6705

+++ Source/Core/Variables.cs
@@ -33,7 +33,7 @@ namespace PHP.Core
 	/// <summary>
 	/// Type codes of Phalanger special variables.
 	/// </summary>
-	public enum PhpTypeCode : short
+	public enum PhpTypeCode : byte
 	{
 		/// <summary>The type code of the <see cref="string"/> type.</summary>
 		String,
